V3.4

+++ Install/one_vpn/add-bird4_routes.sh
@@ -4,14 +4,15 @@
 DEBUG=0
 ISP=ISPINPUT
 VPN1=VPN1INPUT
-HOMEPATH=/opt/root/Bird4Static
-URL0=https://antifilter.download/list/allyouneed.lst
+HOMEPATH=HOMEPATHINPUT
+SYSTEM_FOLDER=SYSTEMFOLDERINPUT
+URL0=URLINPUT
 
  #SCRIPT VARIABE
 BLACKLIST=$HOMEPATH/lists/antifilter.list
-ROUTE_FORCE_ISP=/opt/etc/bird4-force-isp.list
-ROUTE_FORCE_VPN1=/opt/etc/bird4-force-vpn1.list
-ROUTE_BASE_VPN1=/opt/etc/bird4-base-vpn1.list
+ROUTE_FORCE_ISP=$SYSTEM_FOLDER/etc/bird4-force-isp.list
+ROUTE_FORCE_VPN1=$SYSTEM_FOLDER/etc/bird4-force-vpn1.list
+ROUTE_BASE_VPN1=$SYSTEM_FOLDER/etc/bird4-base-vpn1.list
 VPN1TXT=$HOMEPATH/lists/user-vpn.list
 ISPTXT=$HOMEPATH/lists/user-isp.list
 MD5_SUM=$HOMEPATH/scripts/sum.md5
@@ -41,18 +42,21 @@ ipr_func() {
 
  #DIFF FUNCTION
 diff_funk() {
+  if [[ "$3" == "check" ]]; then
+    if grep -q -E "([0-9]{1,3}.){3}[0-9]{1,3}/[0-9]{1,3}" $2; then continue; else return; fi
+  fi
   if [[ "$DEBUG" == 1 ]]; then
     patch_file=/tmp/patch_$(echo $1 | awk -F/ '{print $NF}')
     echo "########### $(date) STEP_3: diff $(echo $1 | awk -F/ '{print $NF}' ) ###########"
     diff -u $1 $2 > $patch_file
-	cat $patch_file && patch $1 $patch_file && rm $patch_file
+    cat $patch_file && patch $1 $patch_file && rm $patch_file
   else
     diff -u $1 $2 | patch $1 -
   fi
 }
 
  #INIT FILES
-if [[ "$DEBUG" == 1 ]]; then echo "$(date) STEP_1: add init files ###########"; fi
+if [[ "$DEBUG" == 1 ]]; then echo "########### $(date) STEP_1: add init files ###########"; fi
 WORK_FILES="$BLACKLIST \
             $ROUTE_FORCE_ISP $ROUTE_FORCE_VPN1 \
             $ROUTE_BASE_VPN1 \
@@ -63,11 +67,11 @@ for var in $WORK_FILES; do
 done
 
  #WAIT DNS
-if [[ "$DEBUG" == 1 ]]; then echo "$(date) STEP_2: wait dns ###########"; fi
+if [[ "$DEBUG" == 1 ]]; then echo "########### $(date) STEP_2: wait dns ###########"; fi
 until ADDRS=$(dig +short google.com @localhost -p 53) && [ -n "$ADDRS" ] > /dev/null 2>&1; do sleep 5; done
 
  #BASE_LIST
-curl -sk $URL0 | sort | diff_funk $BLACKLIST -
+curl -sk $URL0 | sort | diff_funk $BLACKLIST - check
 ipr_func $VPN1 $BLACKLIST | diff_funk $ROUTE_BASE_VPN1 -
 
  #FORCE_LIST
@@ -75,9 +79,9 @@ ipr_func $ISP $ISPTXT | diff_funk $ROUTE_FORCE_ISP -
 ipr_func $VPN1 $VPN1TXT | diff_funk $ROUTE_FORCE_VPN1 -
 
  #RESTART BIRD
-if [[ "$DEBUG" == 1 ]]; then echo "$(date) STEP_4:  restart bird ###########"; fi
-if [ "$(cat $MD5_SUM)" != "$(md5sum /opt/etc/bird4*)" ]; then
-  md5sum /opt/etc/bird4* > $MD5_SUM
+if [[ "$DEBUG" == 1 ]]; then echo "########### $(date) STEP_4: restart bird ###########"; fi
+if [ "$(cat $MD5_SUM)" != "$(md5sum $SYSTEM_FOLDER/etc/bird4*)" ]; then
+  md5sum $SYSTEM_FOLDER/etc/bird4* > $MD5_SUM
   echo "Restarting bird"
   killall -s SIGHUP bird4
-fi
+fi
